https://online.sbis.ru/opendoc.html?guid=00230e73-e840-4986-bcf3-a8362997e300 поправил экспорты и тайпдефы
